fix: remote sdk namespace typing
This changes the default values for some namespace parameters in the
remtoe SDK from None to [], to match the underlying code it calls.

Prior to this commit, failing to supply "namespace" with the remote SDK
would cause an error because the underlying code it dispatches to does
not consider None to be valid input.